찜 기능 새로운 컨트롤러로 분리
